Train on a subset of seasons to avoid overfitting
In order to avoid overfitting a particular game or team within one season, use entire seasons for cross validation. Need to look into downsides of this, but it will avoid situations where the nn has learned features that persist over an entire game, such as over/under, spread, hometeam, and so on.